[FEAT/#323] 홈화면에서 사용할 월별 일기, 일일 일기에 대한 data class를 정의합니다.
