+++ app/src/main/java/com/sopt/clody/domain/model/DailyDiaryInfo.kt
@@ -1,0 +1,6 @@
+package com.sopt.clody.domain.model
+
+data class DailyDiaryInfo(
+    val diaryList: List<String> = listOf(),
+    val isDraft: Boolean = false,
+)
